+++ server/jetstream_cluster_4_test.go
@@ -1,0 +1,950 @@
+// Copyright 2022-2024 The NATS Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build !skip_js_tests && !skip_js_cluster_tests_4
+// +build !skip_js_tests,!skip_js_cluster_tests_4
+
+package server
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"math/rand"
+	"os"
+	"path/filepath"
+	"strings"
+	"sync"
+	"testing"
+	"time"
+
+	"github.com/nats-io/nats.go"
+	"github.com/nats-io/nuid"
+)
+
+func TestJetStreamClusterWorkQueueStreamDiscardNewDesync(t *testing.T) {
+	t.Run("max msgs", func(t *testing.T) {
+		testJetStreamClusterWorkQueueStreamDiscardNewDesync(t, &nats.StreamConfig{
+			Name:      "WQTEST_MM",
+			Subjects:  []string{"messages.*"},
+			Replicas:  3,
+			MaxAge:    10 * time.Minute,
+			MaxMsgs:   100,
+			Retention: nats.WorkQueuePolicy,
+			Discard:   nats.DiscardNew,
+		})
+	})
+	t.Run("max bytes", func(t *testing.T) {
+		testJetStreamClusterWorkQueueStreamDiscardNewDesync(t, &nats.StreamConfig{
+			Name:      "WQTEST_MB",
+			Subjects:  []string{"messages.*"},
+			Replicas:  3,
+			MaxAge:    10 * time.Minute,
+			MaxBytes:  1 * 1024 * 1024,
+			Retention: nats.WorkQueuePolicy,
+			Discard:   nats.DiscardNew,
+		})
+	})
+}
+
+func testJetStreamClusterWorkQueueStreamDiscardNewDesync(t *testing.T, sc *nats.StreamConfig) {
+	conf := `
+	listen: 127.0.0.1:-1
+	server_name: %s
+	jetstream: {
+		store_dir: '%s',
+	}
+	cluster {
+		name: %s
+		listen: 127.0.0.1:%d
+		routes = [%s]
+	}
+        system_account: sys
+        no_auth_user: js
+	accounts {
+	  sys {
+	    users = [
+	      { user: sys, pass: sys }
+	    ]
+	  }
+	  js {
+	    jetstream = enabled
+	    users = [
+	      { user: js, pass: js }
+	    ]
+	  }
+	}`
+	c := createJetStreamClusterWithTemplate(t, conf, sc.Name, 3)
+	defer c.shutdown()
+
+	nc, js := jsClientConnect(t, c.randomServer())
+	defer nc.Close()
+
+	cnc, cjs := jsClientConnect(t, c.randomServer())
+	defer cnc.Close()
+
+	_, err := js.AddStream(sc)
+	require_NoError(t, err)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	psub, err := cjs.PullSubscribe("messages.*", "consumer")
+	require_NoError(t, err)
+
+	stepDown := func() {
+		_, err = nc.Request(fmt.Sprintf(JSApiStreamLeaderStepDownT, sc.Name), nil, time.Second)
+	}
+
+	// Messages will be produced and consumed in parallel, then once there are
+	// enough errors a leader election will be triggered.
+	var (
+		wg          sync.WaitGroup
+		received    uint64
+		errCh       = make(chan error, 100_000)
+		receivedMap = make(map[string]*nats.Msg)
+	)
+	wg.Add(1)
+	go func() {
+		tick := time.NewTicker(20 * time.Millisecond)
+		for {
+			select {
+			case <-ctx.Done():
+				wg.Done()
+				return
+			case <-tick.C:
+				msgs, err := psub.Fetch(10, nats.MaxWait(200*time.Millisecond))
+				if err != nil {
+					// The consumer will continue to timeout here eventually.
+					continue
+				}
+				for _, msg := range msgs {
+					received++
+					receivedMap[msg.Subject] = msg
+					msg.Ack()
+				}
+			}
+		}
+	}()
+
+	shouldDrop := make(map[string]error)
+	wg.Add(1)
+	go func() {
+		payload := []byte(strings.Repeat("A", 1024))
+		tick := time.NewTicker(1 * time.Millisecond)
+		for i := 1; ; i++ {
+			select {
+			case <-ctx.Done():
+				wg.Done()
+				return
+			case <-tick.C:
+				subject := fmt.Sprintf("messages.%d", i)
+				_, err := js.Publish(subject, payload, nats.RetryAttempts(0))
+				if err != nil {
+					errCh <- err
+				}
+				// Capture the messages that have failed.
+				if err != nil {
+					shouldDrop[subject] = err
+				}
+			}
+		}
+	}()
+
+	// Collect enough errors to cause things to get out of sync.
+	var errCount int
+Setup:
+	for {
+		select {
+		case err = <-errCh:
+			errCount++
+			if errCount%500 == 0 {
+				stepDown()
+			} else if errCount >= 2000 {
+				// Stop both producing and consuming.
+				cancel()
+				break Setup
+			}
+		case <-time.After(5 * time.Second):
+			// Unblock the test and continue.
+			cancel()
+			break Setup
+		}
+	}
+
+	// Both goroutines should be exiting now..
+	wg.Wait()
+
+	// Let acks propagate for stream checks.
+	time.Sleep(250 * time.Millisecond)
+
+	// Check messages that ought to have been dropped.
+	for subject := range receivedMap {
+		found, ok := shouldDrop[subject]
+		if ok {
+			t.Errorf("Should have dropped message published on %q since got error: %v", subject, found)
+		}
+	}
+}
+
+// https://github.com/nats-io/nats-server/issues/5071
+func TestJetStreamClusterStreamPlacementDistribution(t *testing.T) {
+	c := createJetStreamClusterExplicit(t, "R3S", 5)
+	defer c.shutdown()
+
+	s := c.randomNonLeader()
+	nc, js := jsClientConnect(t, s)
+	defer nc.Close()
+
+	for i := 1; i <= 10; i++ {
+		_, err := js.AddStream(&nats.StreamConfig{
+			Name:     fmt.Sprintf("TEST:%d", i),
+			Subjects: []string{fmt.Sprintf("foo.%d.*", i)},
+			Replicas: 3,
+		})
+		require_NoError(t, err)
+	}
+
+	// 10 streams, 3 replicas div 5 servers.
+	expectedStreams := 10 * 3 / 5
+	for _, s := range c.servers {
+		jsz, err := s.Jsz(nil)
+		require_NoError(t, err)
+		require_Equal(t, jsz.Streams, expectedStreams)
+	}
+}
+
+func TestJetStreamClusterSourceWorkingQueueWithLimit(t *testing.T) {
+	c := createJetStreamClusterExplicit(t, "WQ3", 3)
+	defer c.shutdown()
+
+	nc, js := jsClientConnect(t, c.randomServer())
+	defer nc.Close()
+
+	_, err := js.AddStream(&nats.StreamConfig{Name: "test", Subjects: []string{"test"}, Replicas: 3})
+	require_NoError(t, err)
+
+	_, err = js.AddStream(&nats.StreamConfig{Name: "wq", MaxMsgs: 100, Discard: nats.DiscardNew, Retention: nats.WorkQueuePolicy,
+		Sources: []*nats.StreamSource{{Name: "test"}}, Replicas: 3})
+	require_NoError(t, err)
+
+	sendBatch := func(subject string, n int) {
+		for i := 0; i < n; i++ {
+			_, err = js.Publish(subject, []byte("OK"))
+			require_NoError(t, err)
+		}
+	}
+	// Populate each one.
+	sendBatch("test", 300)
+
+	checkFor(t, 3*time.Second, 250*time.Millisecond, func() error {
+		si, err := js.StreamInfo("wq")
+		require_NoError(t, err)
+		if si.State.Msgs != 100 {
+			return fmt.Errorf("Expected 100 msgs, got state: %+v", si.State)
+		}
+		return nil
+	})
+
+	_, err = js.AddConsumer("wq", &nats.ConsumerConfig{Durable: "wqc", FilterSubject: "test", AckPolicy: nats.AckExplicitPolicy})
+	require_NoError(t, err)
+
+	ss, err := js.PullSubscribe("test", "wqc", nats.Bind("wq", "wqc"))
+	require_NoError(t, err)
+	// we must have at least one message on the transformed subject name (ie no timeout)
+	f := func(done chan bool) {
+		for i := 0; i < 300; i++ {
+			m, err := ss.Fetch(1, nats.MaxWait(3*time.Second))
+			require_NoError(t, err)
+			time.Sleep(11 * time.Millisecond)
+			err = m[0].Ack()
+			require_NoError(t, err)
+		}
+		done <- true
+	}
+
+	var doneChan = make(chan bool)
+	go f(doneChan)
+
+	checkFor(t, 6*time.Second, 100*time.Millisecond, func() error {
+		si, err := js.StreamInfo("wq")
+		require_NoError(t, err)
+		if si.State.Msgs > 0 && si.State.Msgs <= 100 {
+			return fmt.Errorf("Expected 0 msgs, got: %d", si.State.Msgs)
+		} else if si.State.Msgs > 100 {
+			t.Fatalf("Got more than our 100 message limit: %+v", si.State)
+		}
+		return nil
+	})
+
+	select {
+	case <-doneChan:
+		ss.Drain()
+	case <-time.After(5 * time.Second):
+		t.Fatalf("Did not receive completion signal")
+	}
+}
+
+func TestJetStreamClusterStreamOrphanMsgsAndReplicasDrifting(t *testing.T) {
+	type testParams struct {
+		restartAny       bool
+		restartLeader    bool
+		rolloutRestart   bool
+		ldmRestart       bool
+		restarts         int
+		checkHealthz     bool
+		reconnectRoutes  bool
+		reconnectClients bool
+	}
+	test := func(t *testing.T, params *testParams, sc *nats.StreamConfig) {
+		conf := `
+		listen: 127.0.0.1:-1
+		server_name: %s
+		jetstream: {
+			store_dir: '%s',
+		}
+		cluster {
+			name: %s
+			listen: 127.0.0.1:%d
+			routes = [%s]
+		}
+		server_tags: ["test"]
+		system_account: sys
+		no_auth_user: js
+		accounts {
+			sys { users = [ { user: sys, pass: sys } ] }
+			js {
+				jetstream = enabled
+				users = [ { user: js, pass: js } ]
+		    }
+		}`
+		c := createJetStreamClusterWithTemplate(t, conf, sc.Name, 3)
+		defer c.shutdown()
+
+		// Update lame duck duration for all servers.
+		for _, s := range c.servers {
+			s.optsMu.Lock()
+			s.opts.LameDuckDuration = 5 * time.Second
+			s.opts.LameDuckGracePeriod = -5 * time.Second
+			s.optsMu.Unlock()
+		}
+
+		nc, js := jsClientConnect(t, c.randomServer())
+		defer nc.Close()
+
+		cnc, cjs := jsClientConnect(t, c.randomServer())
+		defer cnc.Close()
+
+		_, err := js.AddStream(sc)
+		require_NoError(t, err)
+
+		pctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+		defer cancel()
+
+		// Start producers
+		var wg sync.WaitGroup
+
+		// First call is just to create the pull subscribers.
+		mp := nats.MaxAckPending(10000)
+		mw := nats.PullMaxWaiting(1000)
+		aw := nats.AckWait(5 * time.Second)
+
+		for i := 0; i < 10; i++ {
+			for _, partition := range []string{"EEEEE"} {
+				subject := fmt.Sprintf("MSGS.%s.*.H.100XY.*.*.WQ.00000000000%d", partition, i)
+				consumer := fmt.Sprintf("consumer:%s:%d", partition, i)
+				_, err := cjs.PullSubscribe(subject, consumer, mp, mw, aw)
+				require_NoError(t, err)
+			}
+		}
+
+		// Create a single consumer that does no activity.
+		// Make sure we still calculate low ack properly and cleanup etc.
+		_, err = cjs.PullSubscribe("MSGS.ZZ.>", "consumer:ZZ:0", mp, mw, aw)
+		require_NoError(t, err)
+
+		subjects := []string{
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000000",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000001",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000002",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000003",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000004",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000005",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000006",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000007",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000008",
+			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000009",
+		}
+		payload := []byte(strings.Repeat("A", 1024))
+
+		for i := 0; i < 50; i++ {
+			wg.Add(1)
+			go func() {
+				pnc, pjs := jsClientConnect(t, c.randomServer())
+				defer pnc.Close()
+
+				for i := 1; i < 200_000; i++ {
+					select {
+					case <-pctx.Done():
+						wg.Done()
+						return
+					default:
+					}
+					for _, subject := range subjects {
+						// Send each message a few times.
+						msgID := nats.MsgId(nuid.Next())
+						pjs.PublishAsync(subject, payload, msgID)
+						pjs.Publish(subject, payload, msgID, nats.AckWait(250*time.Millisecond))
+						pjs.Publish(subject, payload, msgID, nats.AckWait(250*time.Millisecond))
+					}
+				}
+			}()
+		}
+
+		// Rogue publisher that sends the same msg ID everytime.
+		for i := 0; i < 10; i++ {
+			wg.Add(1)
+			go func() {
+				pnc, pjs := jsClientConnect(t, c.randomServer())
+				defer pnc.Close()
+
+				msgID := nats.MsgId("1234567890")
+				for i := 1; ; i++ {
+					select {
+					case <-pctx.Done():
+						wg.Done()
+						return
+					default:
+					}
+					for _, subject := range subjects {
+						// Send each message a few times.
+						pjs.PublishAsync(subject, payload, msgID, nats.RetryAttempts(0), nats.RetryWait(0))
+						pjs.Publish(subject, payload, msgID, nats.AckWait(1*time.Millisecond), nats.RetryAttempts(0), nats.RetryWait(0))
+						pjs.Publish(subject, payload, msgID, nats.AckWait(1*time.Millisecond), nats.RetryAttempts(0), nats.RetryWait(0))
+					}
+				}
+			}()
+		}
+
+		// Let enough messages into the stream then start consumers.
+		time.Sleep(15 * time.Second)
+
+		ctx, cancel := context.WithTimeout(context.Background(), 45*time.Second)
+		defer cancel()
+
+		for i := 0; i < 10; i++ {
+			subject := fmt.Sprintf("MSGS.EEEEE.*.H.100XY.*.*.WQ.00000000000%d", i)
+			consumer := fmt.Sprintf("consumer:EEEEE:%d", i)
+			for n := 0; n < 5; n++ {
+				cpnc, cpjs := jsClientConnect(t, c.randomServer())
+				defer cpnc.Close()
+
+				psub, err := cpjs.PullSubscribe(subject, consumer, mp, mw, aw)
+				require_NoError(t, err)
+
+				time.AfterFunc(15*time.Second, func() {
+					cpnc.Close()
+				})
+
+				wg.Add(1)
+				go func() {
+					tick := time.NewTicker(1 * time.Millisecond)
+					for {
+						if cpnc.IsClosed() {
+							wg.Done()
+							return
+						}
+						select {
+						case <-ctx.Done():
+							wg.Done()
+							return
+						case <-tick.C:
+							// Fetch 1 first, then if no errors Fetch 100.
+							msgs, err := psub.Fetch(1, nats.MaxWait(200*time.Millisecond))
+							if err != nil {
+								continue
+							}
+							for _, msg := range msgs {
+								msg.Ack()
+							}
+							msgs, err = psub.Fetch(100, nats.MaxWait(200*time.Millisecond))
+							if err != nil {
+								continue
+							}
+							for _, msg := range msgs {
+								msg.Ack()
+							}
+							msgs, err = psub.Fetch(1000, nats.MaxWait(200*time.Millisecond))
+							if err != nil {
+								continue
+							}
+							for _, msg := range msgs {
+								msg.Ack()
+							}
+						}
+					}
+				}()
+			}
+		}
+
+		for i := 0; i < 10; i++ {
+			subject := fmt.Sprintf("MSGS.EEEEE.*.H.100XY.*.*.WQ.00000000000%d", i)
+			consumer := fmt.Sprintf("consumer:EEEEE:%d", i)
+			for n := 0; n < 10; n++ {
+				cpnc, cpjs := jsClientConnect(t, c.randomServer())
+				defer cpnc.Close()
+
+				psub, err := cpjs.PullSubscribe(subject, consumer, mp, mw, aw)
+				if err != nil {
+					t.Logf("ERROR: %v", err)
+					continue
+				}
+
+				wg.Add(1)
+				go func() {
+					tick := time.NewTicker(1 * time.Millisecond)
+					for {
+						select {
+						case <-ctx.Done():
+							wg.Done()
+							return
+						case <-tick.C:
+							// Fetch 1 first, then if no errors Fetch 100.
+							msgs, err := psub.Fetch(1, nats.MaxWait(200*time.Millisecond))
+							if err != nil {
+								continue
+							}
+							for _, msg := range msgs {
+								msg.Ack()
+							}
+							msgs, err = psub.Fetch(100, nats.MaxWait(200*time.Millisecond))
+							if err != nil {
+								continue
+							}
+							for _, msg := range msgs {
+								msg.Ack()
+							}
+
+							msgs, err = psub.Fetch(1000, nats.MaxWait(200*time.Millisecond))
+							if err != nil {
+								continue
+							}
+							for _, msg := range msgs {
+								msg.Ack()
+							}
+						}
+					}
+				}()
+			}
+		}
+
+		// Periodically disconnect routes from one of the servers.
+		if params.reconnectRoutes {
+			wg.Add(1)
+			go func() {
+				for range time.NewTicker(10 * time.Second).C {
+					select {
+					case <-ctx.Done():
+						wg.Done()
+						return
+					default:
+					}
+
+					// Force disconnecting routes from one of the servers.
+					s := c.servers[rand.Intn(3)]
+					var routes []*client
+					t.Logf("Disconnecting routes from %v", s.Name())
+					s.mu.Lock()
+					for _, conns := range s.routes {
+						routes = append(routes, conns...)
+					}
+					s.mu.Unlock()
+					for _, r := range routes {
+						r.closeConnection(ClientClosed)
+					}
+				}
+			}()
+		}
+
+		// Periodically reconnect clients.
+		if params.reconnectClients {
+			reconnectClients := func(s *Server) {
+				for _, client := range s.clients {
+					client.closeConnection(Kicked)
+				}
+			}
+
+			wg.Add(1)
+			go func() {
+				for range time.NewTicker(10 * time.Second).C {
+					select {
+					case <-ctx.Done():
+						wg.Done()
+						return
+					default:
+					}
+					// Force reconnect clients from one of the servers.
+					s := c.servers[rand.Intn(len(c.servers))]
+					reconnectClients(s)
+				}
+			}()
+		}
+
+		// Restarts
+		time.AfterFunc(10*time.Second, func() {
+			for i := 0; i < params.restarts; i++ {
+				switch {
+				case params.restartLeader:
+					// Find server leader of the stream and restart it.
+					s := c.streamLeader("js", sc.Name)
+					if params.ldmRestart {
+						s.lameDuckMode()
+					} else {
+						s.Shutdown()
+					}
+					s.WaitForShutdown()
+					c.restartServer(s)
+				case params.restartAny:
+					s := c.servers[rand.Intn(len(c.servers))]
+					if params.ldmRestart {
+						s.lameDuckMode()
+					} else {
+						s.Shutdown()
+					}
+					s.WaitForShutdown()
+					c.restartServer(s)
+				case params.rolloutRestart:
+					for _, s := range c.servers {
+						if params.ldmRestart {
+							s.lameDuckMode()
+						} else {
+							s.Shutdown()
+						}
+						s.WaitForShutdown()
+						c.restartServer(s)
+
+						if params.checkHealthz {
+							hctx, hcancel := context.WithTimeout(ctx, 15*time.Second)
+							defer hcancel()
+
+							for range time.NewTicker(2 * time.Second).C {
+								select {
+								case <-hctx.Done():
+								default:
+								}
+
+								status := s.healthz(nil)
+								if status.StatusCode == 200 {
+									break
+								}
+							}
+						}
+					}
+				}
+				c.waitOnClusterReady()
+			}
+		})
+
+		// Wait until context is done then check state.
+		<-ctx.Done()
+
+		var consumerPending int
+		for i := 0; i < 10; i++ {
+			ci, err := js.ConsumerInfo(sc.Name, fmt.Sprintf("consumer:EEEEE:%d", i))
+			require_NoError(t, err)
+			consumerPending += int(ci.NumPending)
+		}
+
+		getStreamDetails := func(t *testing.T, srv *Server) *StreamDetail {
+			t.Helper()
+			jsz, err := srv.Jsz(&JSzOptions{Accounts: true, Streams: true, Consumer: true})
+			require_NoError(t, err)
+			if len(jsz.AccountDetails) > 0 && len(jsz.AccountDetails[0].Streams) > 0 {
+				stream := jsz.AccountDetails[0].Streams[0]
+				return &stream
+			}
+			t.Error("Could not find account details")
+			return nil
+		}
+
+		checkState := func(t *testing.T) error {
+			t.Helper()
+
+			leaderSrv := c.streamLeader("js", sc.Name)
+			if leaderSrv == nil {
+				return fmt.Errorf("no leader found for stream")
+			}
+			streamLeader := getStreamDetails(t, leaderSrv)
+			var errs []error
+			for _, srv := range c.servers {
+				if srv == leaderSrv {
+					// Skip self
+					continue
+				}
+				stream := getStreamDetails(t, srv)
+				if stream == nil {
+					return fmt.Errorf("stream not found")
+				}
+
+				if stream.State.Msgs != streamLeader.State.Msgs {
+					err := fmt.Errorf("Leader %v has %d messages, Follower %v has %d messages",
+						stream.Cluster.Leader, streamLeader.State.Msgs,
+						srv, stream.State.Msgs,
+					)
+					errs = append(errs, err)
+				}
+				if stream.State.FirstSeq != streamLeader.State.FirstSeq {
+					err := fmt.Errorf("Leader %v FirstSeq is %d, Follower %v is at %d",
+						stream.Cluster.Leader, streamLeader.State.FirstSeq,
+						srv, stream.State.FirstSeq,
+					)
+					errs = append(errs, err)
+				}
+				if stream.State.LastSeq != streamLeader.State.LastSeq {
+					err := fmt.Errorf("Leader %v LastSeq is %d, Follower %v is at %d",
+						stream.Cluster.Leader, streamLeader.State.LastSeq,
+						srv, stream.State.LastSeq,
+					)
+					errs = append(errs, err)
+				}
+			}
+			if len(errs) > 0 {
+				return errors.Join(errs...)
+			}
+			return nil
+		}
+
+		checkMsgsEqual := func(t *testing.T) {
+			// These have already been checked to be the same for all streams.
+			state := getStreamDetails(t, c.streamLeader("js", sc.Name)).State
+			// Gather all the streams.
+			var msets []*stream
+			for _, s := range c.servers {
+				acc, err := s.LookupAccount("js")
+				require_NoError(t, err)
+				mset, err := acc.lookupStream(sc.Name)
+				require_NoError(t, err)
+				msets = append(msets, mset)
+			}
+			for seq := state.FirstSeq; seq <= state.LastSeq; seq++ {
+				var msgId string
+				var smv StoreMsg
+				for _, mset := range msets {
+					mset.mu.RLock()
+					sm, err := mset.store.LoadMsg(seq, &smv)
+					mset.mu.RUnlock()
+					require_NoError(t, err)
+					if msgId == _EMPTY_ {
+						msgId = string(sm.hdr)
+					} else if msgId != string(sm.hdr) {
+						t.Fatalf("MsgIds do not match for seq %d: %q vs %q", seq, msgId, sm.hdr)
+					}
+				}
+			}
+		}
+
+		// Check state of streams and consumers.
+		si, err := js.StreamInfo(sc.Name)
+		require_NoError(t, err)
+
+		// Only check if there are any pending messages.
+		if consumerPending > 0 {
+			streamPending := int(si.State.Msgs)
+			if streamPending != consumerPending {
+				t.Errorf("Unexpected number of pending messages, stream=%d, consumers=%d", streamPending, consumerPending)
+			}
+		}
+
+		// If clustered, check whether leader and followers have drifted.
+		if sc.Replicas > 1 {
+			// If we have drifted do not have to wait too long, usually its stuck for good.
+			checkFor(t, time.Minute, time.Second, func() error {
+				return checkState(t)
+			})
+			// If we succeeded now let's check that all messages are also the same.
+			// We may have no messages but for tests that do we make sure each msg is the same
+			// across all replicas.
+			checkMsgsEqual(t)
+		}
+
+		wg.Wait()
+	}
+
+	// Setting up test variations below:
+	//
+	// File based with single replica and discard old policy.
+	t.Run("R1F", func(t *testing.T) {
+		params := &testParams{
+			restartAny:     true,
+			ldmRestart:     false,
+			rolloutRestart: false,
+			restarts:       1,
+		}
+		test(t, params, &nats.StreamConfig{
+			Name:        "OWQTEST_R1F",
+			Subjects:    []string{"MSGS.>"},
+			Replicas:    1,
+			MaxAge:      30 * time.Minute,
+			Duplicates:  5 * time.Minute,
+			Retention:   nats.WorkQueuePolicy,
+			Discard:     nats.DiscardOld,
+			AllowRollup: true,
+			Placement: &nats.Placement{
+				Tags: []string{"test"},
+			},
+		})
+	})
+
+	// Clustered memory based with discard new policy and max msgs limit.
+	t.Run("R3M", func(t *testing.T) {
+		params := &testParams{
+			restartAny:     true,
+			ldmRestart:     true,
+			rolloutRestart: false,
+			restarts:       1,
+			checkHealthz:   false,
+		}
+		test(t, params, &nats.StreamConfig{
+			Name:        "OWQTEST_R3M",
+			Subjects:    []string{"MSGS.>"},
+			Replicas:    3,
+			MaxAge:      30 * time.Minute,
+			MaxMsgs:     100_000,
+			Duplicates:  5 * time.Minute,
+			Retention:   nats.WorkQueuePolicy,
+			Discard:     nats.DiscardNew,
+			AllowRollup: true,
+			Storage:     nats.MemoryStorage,
+			Placement: &nats.Placement{
+				Tags: []string{"test"},
+			},
+		})
+	})
+
+	// Clustered file based with discard new policy and max msgs limit.
+	t.Run("R3F_DN", func(t *testing.T) {
+		params := &testParams{
+			restartAny:     true,
+			ldmRestart:     true,
+			rolloutRestart: false,
+			restarts:       1,
+		}
+		test(t, params, &nats.StreamConfig{
+			Name:        "OWQTEST_R3F_DN",
+			Subjects:    []string{"MSGS.>"},
+			Replicas:    3,
+			MaxAge:      30 * time.Minute,
+			MaxMsgs:     100_000,
+			Duplicates:  5 * time.Minute,
+			Retention:   nats.WorkQueuePolicy,
+			Discard:     nats.DiscardNew,
+			AllowRollup: true,
+			Placement: &nats.Placement{
+				Tags: []string{"test"},
+			},
+		})
+	})
+
+	// Clustered file based with discard old policy and max msgs limit.
+	t.Run("R3F_DO", func(t *testing.T) {
+		params := &testParams{
+			restartAny:     true,
+			ldmRestart:     true,
+			rolloutRestart: false,
+			restarts:       1,
+		}
+		test(t, params, &nats.StreamConfig{
+			Name:        "OWQTEST_R3F_DO",
+			Subjects:    []string{"MSGS.>"},
+			Replicas:    3,
+			MaxAge:      30 * time.Minute,
+			MaxMsgs:     100_000,
+			Duplicates:  5 * time.Minute,
+			Retention:   nats.WorkQueuePolicy,
+			Discard:     nats.DiscardOld,
+			AllowRollup: true,
+			Placement: &nats.Placement{
+				Tags: []string{"test"},
+			},
+		})
+	})
+
+	// Clustered file based with discard old policy and no limits.
+	t.Run("R3F_DO_NOLIMIT", func(t *testing.T) {
+		params := &testParams{
+			restartAny:       false,
+			ldmRestart:       true,
+			rolloutRestart:   true,
+			restarts:         3,
+			checkHealthz:     true,
+			reconnectRoutes:  true,
+			reconnectClients: true,
+		}
+		test(t, params, &nats.StreamConfig{
+			Name:       "OWQTEST_R3F_DO_NOLIMIT",
+			Subjects:   []string{"MSGS.>"},
+			Replicas:   3,
+			Duplicates: 30 * time.Second,
+			Discard:    nats.DiscardOld,
+			Placement: &nats.Placement{
+				Tags: []string{"test"},
+			},
+		})
+	})
+}
+
+func TestJetStreamClusterConsumerNRGCleanup(t *testing.T) {
+	c := createJetStreamClusterExplicit(t, "R3S", 3)
+	defer c.shutdown()
+
+	nc, js := jsClientConnect(t, c.randomServer())
+	defer nc.Close()
+
+	_, err := js.AddStream(&nats.StreamConfig{
+		Name:      "TEST",
+		Subjects:  []string{"foo"},
+		Storage:   nats.MemoryStorage,
+		Retention: nats.WorkQueuePolicy,
+		Replicas:  3,
+	})
+	require_NoError(t, err)
+
+	// First call is just to create the pull subscribers.
+	_, err = js.PullSubscribe("foo", "dlc")
+	require_NoError(t, err)
+
+	require_NoError(t, js.DeleteConsumer("TEST", "dlc"))
+
+	// Now delete the stream.
+	require_NoError(t, js.DeleteStream("TEST"))
+
+	// Now make sure we cleaned up the NRG directories for the stream and consumer.
+	var numConsumers, numStreams int
+	for _, s := range c.servers {
+		sd := s.JetStreamConfig().StoreDir
+		nd := filepath.Join(sd, "$SYS", "_js_")
+		f, err := os.Open(nd)
+		require_NoError(t, err)
+		dirs, err := f.ReadDir(-1)
+		require_NoError(t, err)
+		for _, fi := range dirs {
+			if strings.HasPrefix(fi.Name(), "C-") {
+				numConsumers++
+			} else if strings.HasPrefix(fi.Name(), "S-") {
+				numStreams++
+			}
+		}
+	}
+	require_Equal(t, numConsumers, 0)
+	require_Equal(t, numStreams, 0)
+}
